multi-stream-vision-app/dashboard/requirements.txt
streamlit
pika
pillow
pandas
numpy
psycopg2-binary
